Allow for future where speed_t maybe unsigned long long

+++ linux-serial-test.c
@@ -21,7 +21,7 @@
 #include <stdbool.h>
 #include <locale.h>	    /* For numeric grouping commas to mark thousands */
 #include <assert.h>	    /* For sanity checking */
-#include <asm-generic/termbits-common.h> /* For speed_t, BOTHER baudrate  */
+#include <asm-generic/termbits-common.h> /* For speed_t (BOTHER baudrate)  */
 static const speed_t _speed_t_max = -1;	 /* typically unsigned int */
 
 #include "setbaudrate.h"	/* For set_custom_baud() via termios2 */
@@ -501,7 +501,7 @@ static void process_options(int argc, char * argv[])
 			double f = atof(optarg);
 			if (f > _speed_t_max || f < 0) {
 				fprintf(stderr, "ERROR: Invalid baud rate %'.0f ", f);
-				fprintf(stderr, "(termios2 max is %'llu)\n", _speed_t_max);
+				fprintf(stderr, "(termios2 max is %'llu)\n", (long long unsigned int) _speed_t_max);
 				exit(-EINVAL);
 			}
 			_cl_baud = (int)atof(optarg); 	/* Allow 3E6 instead of 3000000 */
